requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrdict==2.0.1
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
bce-python-sdk==0.9.6
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cssselect==1.2.0
cssutils==2.10.2
cycler==0.12.1
Cython==3.0.10
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.29.2
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
durationpy==0.7
easyocr==1.7.1
emoji==2.13.2
et-xmlfile==1.1.0
executing==2.0.1
faiss-cpu==1.9.0
fastjsonschema==2.19.1
filelock==3.13.4
filetype==1.2.0
fire==0.6.0
Flask==3.0.3
flask-babel==4.0.0
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.70.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
groq==0.11.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.2
idna==3.7
imageio==2.34.0
imgaug==0.4.0
importlib_metadata==8.2.0
importlib_resources==6.4.5
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.24
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.12
langchain-google-genai==2.0.1
langchain-groq==0.2.0
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.136
lazy_loader==0.4
lmdb==1.4.1
looseversion==1.3.0
lxml==5.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.8.2
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
ollama==0.3.0
omegaconf==2.3.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opentelemetry-proto==1.27.0
opentelemetry-util-http==0.48b0
orderly-set==5.2.2
orjson==3.10.7
overrides==7.7.0
packaging==24.0
paddleocr==2.7.3
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pdf2docx==0.5.8
pdf2image==1.17.0
pi_heif==0.18.0
pillow==10.3.0
platformdirs==4.2.0
portalocker==2.10.1
posthog==3.6.6
premailer==3.10.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==5.28.2
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post5
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.17.2
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pypandoc==1.13
pyparsing==3.1.2
pypdf==5.0.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.5.4
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.10
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rapidfuzz==3.8.1
rarfile==4.2
referencing==0.34.0
regex==2024.7.24
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.2
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.3
scikit-image==0.23.1
scikit-learn==1.5.2
scipy==1.13.0
seaborn==0.13.2
Send2Trash==1.8.3
shapely==2.0.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.38.6
streamlit==1.39.0
sympy==1.12
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
terminado==0.18.1
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.2.12
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.43.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.1.46
unstructured.pytesseract==0.3.13
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.6
visualdl==2.5.3
Wand==0.6.13
watchdog==4.0.2
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==13.1
Werkzeug==3.0.2
widgetsnbextension==4.0.10
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.12.1
zipp==3.19.2
